+++ rust/connlib/shared/src/lib.rs
@@ -45,19 +45,6 @@ pub const DEFAULT_MTU: u32 = 1280;
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 const LIB_NAME: &str = "connlib";
 
-/// Deactivates DNS control on Windows
-#[cfg(target_os = "windows")]
-pub fn deactivate_dns_control() -> anyhow::Result<()> {
-    windows::dns::deactivate()
-}
-
-/// Deactivates DNS control on other platforms (does nothing)
-#[cfg(not(target_os = "windows"))]
-#[allow(clippy::unnecessary_wraps)]
-pub fn deactivate_dns_control() -> anyhow::Result<()> {
-    Ok(())
-}
-
 pub fn keypair() -> (StaticSecret, PublicKey) {
     let private_key = StaticSecret::random_from_rng(OsRng);
     let public_key = PublicKey::from(&private_key);
