+++ rust/connlib/tunnel/src/device_channel/tun_windows.rs
@@ -115,7 +115,7 @@ impl Tun {
         }
 
         // TODO: Might be calling this more often than it needs
-        connlib_shared::windows::dns::flush().expect("Should be able to flush Windows' DNS cache");
+        flush_dns().expect("Should be able to flush Windows' DNS cache");
         self.routes = new_routes;
         Ok(())
     }
@@ -221,6 +221,16 @@ impl Tun {
 
         row
     }
+}
+
+/// Flush Windows' system-wide DNS cache
+pub(crate) fn flush_dns() -> Result<()> {
+    tracing::info!("Flushing Windows DNS cache");
+    Command::new("powershell")
+        .creation_flags(CREATE_NO_WINDOW)
+        .args(["-Command", "Clear-DnsClientCache"])
+        .status()?;
+    Ok(())
 }
 
 fn start_recv_thread(
