+++ rust/gui-client/src-tauri/src/client/deep_link/windows.rs
@@ -79,7 +79,7 @@ pub async fn open(url: &url::Url) -> Result<()> {
 }
 
 fn pipe_path() -> String {
-    firezone_headless_client::platform::named_pipe_path(&format!("{BUNDLE_ID}.deep_link"))
+    firezone_headless_client::windows::named_pipe_path(&format!("{BUNDLE_ID}.deep_link"))
 }
 
 /// Registers the current exe as the handler for our deep link scheme.
